Extract talent and location matching into parallel sub-workflows #407
Move character extraction + talent matching and location extraction + location
matching out of the monolithic analyze-script workflow into dedicated
talentMatchingWorkflow and locationMatchingWorkflow. These are invoked in
parallel via context.invoke, reducing sequential LLM calls during analysis.
