+++ src/routes/api/workflows/$.ts
@@ -17,6 +17,7 @@ import { generateImageWorkflow } from '@/lib/workflows/image-workflow';
 import { libraryLocationSheetWorkflow } from '@/lib/workflows/library-location-sheet-workflow';
 import { libraryTalentSheetWorkflow } from '@/lib/workflows/library-talent-sheet-workflow';
 import { locationBibleWorkflow } from '@/lib/workflows/location-bible-workflow';
+import { locationMatchingWorkflow } from '@/lib/workflows/location-matching-workflow';
 import { locationSheetWorkflow } from '@/lib/workflows/location-sheet-workflow';
 import { mergeAudioVideoWorkflow } from '@/lib/workflows/merge-audio-video-workflow';
 import { mergeVideoWorkflow } from '@/lib/workflows/merge-video-workflow';
@@ -28,6 +29,7 @@ import { recastCharacterWorkflow } from '@/lib/workflows/recast-character-workfl
 import { recastLocationWorkflow } from '@/lib/workflows/recast-location-workflow';
 import { regenerateFramesWorkflow } from '@/lib/workflows/regenerate-frames-workflow';
 import { generateStoryboardWorkflow } from '@/lib/workflows/storyboard-workflow';
+import { talentMatchingWorkflow } from '@/lib/workflows/talent-matching-workflow';
 import { upscaleVariantWorkflow } from '@/lib/workflows/upscale-variant-workflow';
 import { generateVariantWorkflow } from '@/lib/workflows/variant-workflow';
 import { visualPromptSceneWorkflow } from '@/lib/workflows/visual-prompt-scene-workflow';
@@ -59,10 +61,12 @@ function getHandler() {
         'upscale-variant': upscaleVariantWorkflow,
         'recast-character': recastCharacterWorkflow,
         'recast-location': recastLocationWorkflow,
+        'location-matching': locationMatchingWorkflow,
         'location-sheet': locationSheetWorkflow,
         'location-sheet-from-bible': locationBibleWorkflow,
         'library-location-sheet': libraryLocationSheetWorkflow,
         'regenerate-frames': regenerateFramesWorkflow,
+        'talent-matching': talentMatchingWorkflow,
         'visual-prompt-scene': visualPromptSceneWorkflow,
         'motion-prompts': motionPromptWorkflow,
         'motion-prompt-scene': motionPromptSceneWorkflow,
